requirements.txt
streamlit 
requests 
numpy 
faiss-cpu
python-dotenv
